farmer: Remove dsn-archiving and refactor dsn-sync.
- remove dsn-achiving code
- move sync code into a separate file and disable it
- refactor tests
